--- src/Portkey.Indexer.CA/Processors/CAHolderCreatedLogEventProcessor.cs
@@ -1,118 +1,0 @@
-using AElfIndexer.Client;
-using AElfIndexer.Client.Handlers;
-using AElfIndexer.Grains.State.Client;
-using Google.Protobuf.WellKnownTypes;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Portkey.Contracts.CA;
-using Portkey.Indexer.CA.Entities;
-using Volo.Abp.ObjectMapping;
-using Guardian = Portkey.Indexer.CA.Entities.Guardian;
-
-namespace Portkey.Indexer.CA.Processors;
-
-public class CAHolderCreatedLogEventProcessor: CAHolderTransactionProcessorBase<CAHolderCreated>
-{
-    
-    public CAHolderCreatedLogEventProcessor(
-        ILogger<CAHolderCreatedLogEventProcessor> logger,
-        IAElfIndexerClientEntityRepository<CAHolderIndex, TransactionInfo> caHolderIndexRepository,
-        IAElfIndexerClientEntityRepository<CAHolderManagerIndex, TransactionInfo> caHolderManagerIndexRepository,
-        IAElfIndexerClientEntityRepository<CAHolderTransactionIndex, TransactionInfo>
-            caHolderTransactionIndexRepository,
-        IAElfIndexerClientEntityRepository<TokenInfoIndex, TransactionInfo> tokenInfoIndexRepository,
-        IAElfIndexerClientEntityRepository<NFTInfoIndex, TransactionInfo> nftInfoIndexRepository,
-        IAElfIndexerClientEntityRepository<CAHolderTransactionAddressIndex, TransactionInfo>
-            caHolderTransactionAddressIndexRepository,
-        IOptionsSnapshot<ContractInfoOptions> contractInfoOptions,
-        IOptionsSnapshot<CAHolderTransactionInfoOptions> caHolderTransactionInfoOptions, IObjectMapper objectMapper) :
-        base(logger, caHolderIndexRepository, caHolderManagerIndexRepository, caHolderTransactionIndexRepository,
-            tokenInfoIndexRepository,
-            nftInfoIndexRepository, caHolderTransactionAddressIndexRepository, contractInfoOptions,
-            caHolderTransactionInfoOptions, objectMapper)
-    {
-    }
-    
-    public override string GetContractAddress(string chainId)
-    {
-        return ContractInfoOptions.ContractInfos.First(c=>c.ChainId == chainId).CAContractAddress;
-    }
-
-    protected override async Task HandleEventAsync(CAHolderCreated eventValue, LogEventContext context)
-    {
-        await HandlerTransactionIndexAsync(eventValue, context);
-        //check manager is already exist in caHolderManagerIndex
-        var managerIndexId = IdGenerateHelper.GetId(context.ChainId, eventValue.Manager.ToBase58());
-        var caHolderManagerIndex =
-            await CAHolderManagerIndexRepository.GetFromBlockStateSetAsync(managerIndexId, context.ChainId);
-        if (caHolderManagerIndex == null)
-        {
-            caHolderManagerIndex = new CAHolderManagerIndex
-            {
-                Id = managerIndexId,
-                Manager = eventValue.Manager.ToBase58(),
-                CAAddresses = new List<string>()
-                {
-                    eventValue.CaAddress.ToBase58()
-                }
-            };
-        }
-        else
-        {
-            if (!caHolderManagerIndex.CAAddresses.Contains(eventValue.CaAddress.ToBase58()))
-            {
-                caHolderManagerIndex.CAAddresses.Add(eventValue.CaAddress.ToBase58());
-            }
-        }
-        ObjectMapper.Map(context, caHolderManagerIndex);
-        await CAHolderManagerIndexRepository.AddOrUpdateAsync(caHolderManagerIndex);
-        
-        
-        //check ca address if already exist in caHolderIndex
-        var indexId = IdGenerateHelper.GetId(context.ChainId, eventValue.CaAddress.ToBase58());
-        var caHolderIndex = await CAHolderIndexRepository.GetFromBlockStateSetAsync(indexId, context.ChainId);
-        if (caHolderIndex != null)
-        {
-            return;
-        }
-        
-        // _objectMapper.Map<LogEventContext, CAHolderIndex>(context, caHolderIndex);
-
-        caHolderIndex = new CAHolderIndex
-        {
-            Id = indexId,
-            CAHash = eventValue.CaHash.ToHex(),
-            CAAddress = eventValue.CaAddress.ToBase58(),
-            Creator = eventValue.Creator.ToBase58(),
-            ManagerInfos = new List<Entities.ManagerInfo>
-            {
-                new ()
-                {
-                    Address = eventValue.Manager.ToBase58(),
-                    ExtraData = eventValue.ExtraData
-                }
-            },
-            Guardians = new List<Guardian>(),
-            OriginChainId = context.ChainId
-        };
-        ObjectMapper.Map(context, caHolderIndex);
-        await CAHolderIndexRepository.AddOrUpdateAsync(caHolderIndex);
-    }
-
-    protected override async Task HandlerTransactionIndexAsync(CAHolderCreated eventValue, LogEventContext context)
-    {
-        if (!IsValidTransaction(context.ChainId, context.To, context.MethodName, context.Params)) return;
-        
-        var index = new CAHolderTransactionIndex
-        {
-            Id = IdGenerateHelper.GetId(context.BlockHash, context.TransactionId),
-            Timestamp = context.BlockTime.ToTimestamp().Seconds,
-            FromAddress = eventValue.CaAddress.ToBase58(),
-            TransactionFee = GetTransactionFee(context.ExtraProperties)
-        };
-        
-        ObjectMapper.Map(context, index);
-        index.MethodName = GetMethodName(context.MethodName, context.Params);
-        await CAHolderTransactionIndexRepository.AddOrUpdateAsync(index);
-    }
-}
