+++ src/Portkey.Indexer.CA/Processors/InvitedProcessor.cs
@@ -10,13 +10,13 @@ using Volo.Abp.ObjectMapping;
 
 namespace Portkey.Indexer.CA.Processors;
 
-public class InviteLogEventProcessor : AElfLogEventProcessorBase<Invited, TransactionInfo>
+public class InvitedProcessor : AElfLogEventProcessorBase<Invited, TransactionInfo>
 {
     private readonly IObjectMapper _objectMapper;
     private readonly IAElfIndexerClientEntityRepository<InviteIndex, TransactionInfo> _repository;
     private readonly ContractInfoOptions _contractInfoOptions;
 
-    public InviteLogEventProcessor(ILogger<InviteLogEventProcessor> logger, IObjectMapper objectMapper,
+    public InvitedProcessor(ILogger<InvitedProcessor> logger, IObjectMapper objectMapper,
         IAElfIndexerClientEntityRepository<InviteIndex, TransactionInfo> repository,
         IOptionsSnapshot<ContractInfoOptions> contractInfoOptions) : base(logger)
     {
