+++ src/Portkey.Indexer.CA/Processors/ManagerUpdatedProcessor.cs
@@ -1,7 +1,6 @@
 using AElfIndexer.Client;
 using AElfIndexer.Client.Handlers;
 using AElfIndexer.Grains.State.Client;
-using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Portkey.Contracts.CA;
@@ -10,24 +9,18 @@ using Volo.Abp.ObjectMapping;
 
 namespace Portkey.Indexer.CA.Processors;
 
-[Obsolete]
-public class ManagerUpdatedProcessor : CAHolderTransactionProcessorBase<ManagerInfoUpdated>
+public class ManagerUpdatedProcessor : CAHolderManagerProcessorBase<ManagerInfoUpdated>
 {
     public ManagerUpdatedProcessor(ILogger<ManagerUpdatedProcessor> logger,
-        IAElfIndexerClientEntityRepository<CAHolderIndex, TransactionInfo> caHolderIndexRepository,
+        IObjectMapper objectMapper, IOptionsSnapshot<ContractInfoOptions> contractInfoOptions,
+        IAElfIndexerClientEntityRepository<CAHolderIndex, TransactionInfo> repository,
         IAElfIndexerClientEntityRepository<CAHolderManagerIndex, TransactionInfo> caHolderManagerIndexRepository,
-        IAElfIndexerClientEntityRepository<CAHolderTransactionIndex, TransactionInfo>
-            caHolderTransactionIndexRepository,
-        IAElfIndexerClientEntityRepository<TokenInfoIndex, TransactionInfo> tokenInfoIndexRepository,
-        IAElfIndexerClientEntityRepository<NFTInfoIndex, TransactionInfo> nftInfoIndexRepository,
-        IAElfIndexerClientEntityRepository<CAHolderTransactionAddressIndex, TransactionInfo>
-            caHolderTransactionAddressIndexRepository,
-        IOptionsSnapshot<ContractInfoOptions> contractInfoOptions,
-        IOptionsSnapshot<CAHolderTransactionInfoOptions> caHolderTransactionInfoOptions, IObjectMapper objectMapper) :
-        base(logger, caHolderIndexRepository, caHolderManagerIndexRepository, caHolderTransactionIndexRepository,
-            tokenInfoIndexRepository,
-            nftInfoIndexRepository, caHolderTransactionAddressIndexRepository, contractInfoOptions,
-            caHolderTransactionInfoOptions, objectMapper)
+        IAElfIndexerClientEntityRepository<CAHolderManagerChangeRecordIndex, TransactionInfo> changeRecordRepository,
+        IAElfIndexerClientEntityRepository<CAHolderTransactionIndex, TransactionInfo> caHolderTransactionIndexRepository,
+        IAElfIndexerClientEntityRepository<CAHolderTransactionAddressIndex, TransactionInfo> caHolderTransactionAddressIndexRepository,
+        IOptionsSnapshot<CAHolderTransactionInfoOptions> caHolderTransactionInfoOptions) :
+        base(logger, objectMapper, contractInfoOptions, repository,caHolderManagerIndexRepository, changeRecordRepository,
+            caHolderTransactionIndexRepository, caHolderTransactionAddressIndexRepository, caHolderTransactionInfoOptions)
     {
     }
 
@@ -37,6 +30,26 @@ public class ManagerUpdatedProcessor : CAHolderTransactionProcessorBase<ManagerI
     }
 
     protected override async Task HandleEventAsync(ManagerInfoUpdated eventValue, LogEventContext context)
+    {
+        await HandlerTransactionIndexAsync(eventValue, context);
+        //check ca address if already exist in caHolderIndex
+        var indexId = IdGenerateHelper.GetId(context.ChainId, eventValue.CaAddress.ToBase58());
+        var caHolderIndex = await Repository.GetFromBlockStateSetAsync(indexId, context.ChainId);
+        if (caHolderIndex == null) return;
+
+        ObjectMapper.Map(context, caHolderIndex);
+
+        var managerInfo = caHolderIndex.ManagerInfos.FirstOrDefault(m => m.Address == eventValue.Manager.ToBase58());
+        if (managerInfo == null) return;
+
+        managerInfo.ExtraData = eventValue.ExtraData;
+
+        await Repository.AddOrUpdateAsync(caHolderIndex);
+        await AddChangeRecordAsync(eventValue.CaAddress.ToBase58(), eventValue.CaHash.ToHex(),
+            eventValue.Manager.ToBase58(), nameof(ManagerInfoUpdated), context);
+    }
+    
+    protected override async Task HandlerTransactionIndexAsync(ManagerInfoUpdated eventValue, LogEventContext context)
     {
         await ProcessCAHolderTransactionAsync(context, eventValue.CaAddress.ToBase58());
     }
