+++ src/Portkey.Indexer.CA/Processors/TokenCrossChainReceivedProcessor.cs
@@ -12,9 +12,9 @@ using Volo.Abp.ObjectMapping;
 
 namespace Portkey.Indexer.CA.Processors;
 
-public class TokenCrossChainReceivedLogEventProcessor:  CAHolderTokenBalanceProcessorBase<CrossChainReceived>
+public class TokenCrossChainReceivedProcessor:  CAHolderTokenBalanceProcessorBase<CrossChainReceived>
 {
-    public TokenCrossChainReceivedLogEventProcessor(ILogger<TokenCrossChainReceivedLogEventProcessor> logger,
+    public TokenCrossChainReceivedProcessor(ILogger<TokenCrossChainReceivedProcessor> logger,
         IOptionsSnapshot<ContractInfoOptions> contractInfoOptions,
         IAElfIndexerClientEntityRepository<CAHolderIndex, TransactionInfo> caHolderIndexRepository,
         IAElfIndexerClientEntityRepository<TokenInfoIndex, TransactionInfo> tokenInfoIndexRepository,
