+++ test/Portkey.Indexer.CA.Tests/Processors/ManagerLogEventProcessorTests.cs
@@ -66,7 +66,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedLogEventProcessor>();
+        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -149,7 +149,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedLogEventProcessor>();
+        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -229,7 +229,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedLogEventProcessor>();
+        var managerAddedLogEventProcessor = GetRequiredService<ManagerAddedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -307,7 +307,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerRemovedLogEventProcessor = GetRequiredService<ManagerRemovedLogEventProcessor>();
+        var managerRemovedLogEventProcessor = GetRequiredService<ManagerRemovedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -386,7 +386,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerRemovedLogEventProcessor = GetRequiredService<ManagerRemovedLogEventProcessor>();
+        var managerRemovedLogEventProcessor = GetRequiredService<ManagerRemovedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -696,7 +696,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerUpdatedLogEventProcessor = GetRequiredService<ManagerUpdatedLogEventProcessor>();
+        var managerUpdatedLogEventProcessor = GetRequiredService<ManagerUpdatedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -777,7 +777,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var managerUpdatedLogEventProcessor = GetRequiredService<ManagerUpdatedLogEventProcessor>();
+        var managerUpdatedLogEventProcessor = GetRequiredService<ManagerUpdatedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -1027,7 +1027,7 @@ public sealed class ManagerLogEventProcessorTests : PortkeyIndexerCATestBase
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
 
-        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedLogEventProcessor>();
+        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
