+++ test/Portkey.Indexer.CA.Tests/Processors/PlayedLogEventProcessorTests.cs
@@ -120,7 +120,7 @@ public class PlayedLogEventProcessorTests: PortkeyIndexerCATestBase
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
 
-        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedLogEventProcessor>();
+        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
