+++ test/Portkey.Indexer.CA.Tests/Processors/TransactionFeeChargedLogEventProcessorTests.cs
@@ -26,7 +26,7 @@ public class TransactionFeeChargedLogEventProcessorTests : PortkeyIndexerCATestB
         const string previousBlockHash = "e38c4fb1cf6af05878657cb3f7b5fc8a5fcfb2eec19cd76b73abb831973fbf4e";
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
-        var transactionFeeChargedLogEventProcessor = GetRequiredService<TransactionFeeChargedLogEventProcessor>();
+        var transactionFeeChargedLogEventProcessor = GetRequiredService<TransactionFeeChargedProcessor>();
         
         
         var blockStateSet = new BlockStateSet<TransactionInfo>
@@ -82,7 +82,7 @@ public class TransactionFeeChargedLogEventProcessorTests : PortkeyIndexerCATestB
         const string transactionId = "c1e625d135171c766999274a00a7003abed24cfe59a7215aabf1472ef20a2da2";
         const long blockHeight = 100;
 
-        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedLogEventProcessor>();
+        var caHolderCreatedProcessor = GetRequiredService<CAHolderCreatedProcessor>();
 
         //step1: create blockStateSet
         var blockStateSet = new BlockStateSet<TransactionInfo>
